split up the old and the new
